+++ reactjs-ecm/src/services/auth-api.js
@@ -8,14 +8,3 @@ export async function loginApi(params) {
     return error?.response.data;
   }
 }
-const API_BASE_URL = 'http://localhost:6006';
-
-export const fetchUsers = async (page, limit) => {
-  try {
-    const response = await axios.get(`${API_BASE_URL}/users/${page}/${limit}`);
-    return response.data;
-  } catch (error) {
-    console.error('Error fetching users:', error);
-    throw error;
-  }
-};
