+++ reactjs-ecm/src/services/custom-auth-api.js
@@ -1,5 +1,5 @@
 import axios from "axios";
-import { fetchUsers as fetchUsersFromAPI } from './auth-api';
+
 export const apiClient = axios.create({
   baseURL: "http://localhost:6006",
   headers: {
@@ -7,8 +7,3 @@ export const apiClient = axios.create({
   },
   timeout: 1000,
 });
-export const fetchUsers = async (page, limit) => {
-  const result = await fetchUsersFromAPI(page, limit);
-  // Tùy chỉnh dữ liệu nếu cần
-  return result;
-};
